+++ mex/editor/merge/state.py
@@ -1,5 +1,5 @@
 from collections.abc import Generator, Iterable
-from typing import Any, Literal
+from typing import Literal
 
 import reflex as rx
 from reflex.event import EventSpec
@@ -62,20 +62,18 @@ class MergeState(State):
         )
 
     @rx.var
-    def label_entity_types_extracted(self) -> list[dict[str, Any]]:
+    def label_entity_types_extracted(self) -> list[ValueLabelCheckboxItem]:
         """Get entity_types_extracted with value, label and checked."""
         return sorted(
             [
-                {
-                    "label": self._locale_service.get_ui_label(
-                        self.current_locale, key
-                    ),
-                    "value": key,
-                    "checked": self.entity_types_extracted[key],
-                }
+                ValueLabelCheckboxItem(
+                    label=self._locale_service.get_ui_label(self.current_locale, key),
+                    value=key,
+                    checked=self.entity_types_extracted[key],
+                )
                 for key in self.entity_types_extracted
             ],
-            key=lambda x: x["label"],
+            key=lambda x: x.label,
         )
 
     @rx.event